test(discord): keep live policy proof at the integration boundary
