+++ src/gateway/server-reload-handlers.test.ts
@@ -46,7 +46,6 @@ import {
   setGatewaySigusr1RestartPolicy,
   setPreRestartDeferralCheck,
 } from "../infra/restart.js";
-import type { fanInChannelIngressLifecycles } from "../plugin-sdk/channel-ingress-runtime.js";
 import { createPluginRuntimeCapabilityLease } from "../plugins/capability-lease.js";
 import { registerPluginCommandInRegistry } from "../plugins/command-registration.js";
 import {
@@ -105,7 +104,6 @@ import {
   type PreparedSecretsRuntimeSnapshot,
 } from "../secrets/runtime.js";
 import { openOpenClawStateDatabase } from "../state/openclaw-state-db.js";
-import { loadBundledPluginFacade } from "../test-utils/bundled-plugin-public-surface.js";
 import { createChannelTestPluginBase, createTestRegistry } from "../test-utils/channel-plugins.js";
 import { isRecord } from "../utils.js";
 import { diffConfigPaths, diffGatewayReloadPaths } from "./config-diff.js";
@@ -4345,207 +4343,6 @@ describe("gateway restart deferral preflight", () => {
     expect(runtimePublished).toBe(true);
     expect(setState).toHaveBeenCalledTimes(1);
     expect(getDeferredChannelReloads?.()).toEqual([]);
-  });
-
-  it("admits newly allowed Discord messages during active work without restarting", async () => {
-    // Describe the public calls this fixture consumes without importing plugin implementation types.
-    const discord = await loadBundledPluginFacade<{
-      createDiscordMessageHandler: (options: ReturnType<typeof createHandlerOptions>) => {
-        (raw: ReturnType<typeof createRawMessage>, transport: typeof client): Promise<void>;
-        deactivate: () => Promise<void>;
-      };
-      createNoopThreadBindingManager: (accountId: string) => object;
-    }>({ pluginId: "discord", artifactBasename: "runtime-api.js" });
-    const { discordPlugin } = await loadBundledPluginFacade<{ discordPlugin: ChannelPlugin }>({
-      pluginId: "discord",
-      artifactBasename: "api.js",
-    });
-    const cfg: OpenClawConfig = {
-      channels: { discord: { token: "synthetic-token", groupPolicy: "allowlist", guilds: {} } },
-      messages: { inbound: { debounceMs: 0 } },
-    };
-    setActivePluginRegistry(
-      createTestRegistry([{ pluginId: "discord", plugin: discordPlugin, source: "synthetic" }]),
-    );
-    setRuntimeConfigSnapshot(cfg, cfg);
-    const restoreChannelReloadEnv = enableChannelReloadsForTest();
-    let channelId = "456";
-    let pendingChannelLookup: Promise<void> | undefined;
-    let channelLookupStarted = false;
-    const guildId = "123";
-    const userId = "789";
-    // The synthetic transport supplies only metadata lookups; preflight stays real.
-    const client = {
-      fetchChannel: async () => {
-        channelLookupStarted = true;
-        await pendingChannelLookup;
-        return { id: channelId, type: 0, name: "synthetic" };
-      },
-      fetchGuild: async () => null,
-    };
-    const admitted: string[] = [];
-    const abort = new AbortController();
-    const createTransportEvent = (raw: ReturnType<typeof createRawMessage>) => ({
-      author: raw.author,
-      guild_id: guildId,
-      guild: { id: guildId, name: "synthetic" },
-      channel_id: channelId,
-      message: {
-        ...raw,
-        channelId,
-        mentionedUsers: [],
-        mentionedRoles: [],
-        mentionedEveryone: false,
-      },
-    });
-    const createHandlerOptions = (threadBindings: object) => ({
-      client,
-      cfg,
-      discordConfig: cfg.channels?.discord,
-      accountId: "default",
-      token: "synthetic-token",
-      runtime: {
-        log: vi.fn(),
-        error: vi.fn(),
-        exit: (code: number) => {
-          throw new Error(`unexpected exit ${code}`);
-        },
-      },
-      botUserId: "999",
-      guildHistories: new Map(),
-      historyLimit: 0,
-      mediaMaxBytes: 1_000,
-      textLimit: 2_000,
-      replyToMode: "off",
-      dmEnabled: true,
-      dmPolicy: "pairing",
-      groupDmEnabled: false,
-      guildEntries: {},
-      threadBindings,
-      abortSignal: abort.signal,
-      testing: {
-        processDiscordMessage: async ({ message }: { message: { id: string } }) => {
-          admitted.push(message.id);
-        },
-        createIngressMonitor: ({
-          dispatch,
-        }: {
-          dispatch: (
-            event: ReturnType<typeof createTransportEvent>,
-            lifecycle: NonNullable<Parameters<typeof fanInChannelIngressLifecycles>[0][number]>,
-          ) => Promise<unknown>;
-        }) => ({
-          start() {},
-          async stop() {},
-          async accept(raw: ReturnType<typeof createRawMessage>) {
-            await dispatch(createTransportEvent(raw), {
-              abortSignal: abort.signal,
-              onAdopted() {},
-              onDeferred() {},
-              onAdoptionFinalizing() {},
-              onAbandoned() {},
-            });
-          },
-        }),
-      },
-    });
-    const createRawMessage = (id: string) => ({
-      id,
-      channel_id: channelId,
-      content: "hello",
-      author: {
-        id: userId,
-        username: "synthetic",
-        discriminator: "0",
-        avatar: null,
-        global_name: null,
-      },
-      attachments: [],
-      embeds: [],
-      mentions: [],
-      mention_roles: [],
-      components: [],
-      mention_everyone: false,
-      timestamp: "2026-09-06T00:00:00.000Z",
-      edited_timestamp: null,
-      type: 0,
-      tts: false,
-      pinned: false,
-    });
-    const handler = discord.createDiscordMessageHandler(
-      createHandlerOptions(discord.createNoopThreadBindingManager("default")),
-    );
-    const send = (id: string) => handler(createRawMessage(id), client);
-    const startChannel = vi.fn(async () => new Map());
-    const stopChannel = vi.fn(async () => {});
-    const { applyHotReload } = createGatewayReloadHandlers({ startChannel, stopChannel });
-    let pendingReload: Promise<unknown> | undefined;
-    let committed = cfg;
-    const apply = async (next: OpenClawConfig) => {
-      const paths = diffGatewayReloadPaths(committed, next, listConfigReloadRefinementPrefixes());
-      const plan = buildGatewayReloadPlan(paths, {
-        previousConfig: committed,
-        candidateConfig: next,
-      });
-      pendingReload = applyHotReload(plan, next, {
-        sourceConfig: next,
-        isCurrent: () => true,
-        publish: async (commit) => {
-          setRuntimeConfigSnapshot(next, next);
-          await commit();
-          committed = next;
-        },
-      });
-      await waitForFast(() =>
-        expect(committed, "policy-only publication must not wait for the active turn").toBe(next),
-      );
-      await pendingReload;
-    };
-    hoisted.activeEmbeddedRunCount.value = 1;
-    try {
-      await send("before-policy");
-      expect(admitted).toEqual([]);
-      await apply({
-        ...cfg,
-        channels: {
-          discord: {
-            ...cfg.channels?.discord,
-            guilds: {
-              [guildId]: { users: [userId], requireMention: false },
-            },
-          },
-        },
-      });
-      await send("newly-allowed");
-      await waitForFast(() => expect(admitted).toEqual(["newly-allowed"]));
-      channelId = "457";
-      channelLookupStarted = false;
-      let releaseChannelLookup!: () => void;
-      pendingChannelLookup = new Promise<void>((resolve) => {
-        releaseChannelLookup = resolve;
-      });
-      const pendingMessage = send("pending-revocation");
-      try {
-        await waitForFast(() => expect(channelLookupStarted).toBe(true));
-        await apply(cfg);
-      } finally {
-        releaseChannelLookup();
-        await pendingMessage;
-      }
-      pendingChannelLookup = undefined;
-      await send("after-revocation");
-      expect(admitted).toEqual(["newly-allowed"]);
-      expect(hoisted.activeEmbeddedRunCount.value).toBe(1);
-      expect(startChannel).not.toHaveBeenCalled();
-      expect(stopChannel).not.toHaveBeenCalled();
-    } finally {
-      hoisted.activeEmbeddedRunCount.value = 0;
-      await pendingReload?.catch(() => {});
-      abort.abort();
-      await handler.deactivate();
-      clearRuntimeConfigSnapshot();
-      restoreChannelReloadEnv();
-    }
   });
 
   it("uses the default channel reload deferral timeout when config omits deferralTimeoutMs", async () => {
